+++ xdef/src/test/java/X.java
@@ -2,11 +2,11 @@ import java.io.StringWriter;
 import org.xdef.XDConstants;
 import org.xdef.XDDocument;
 import org.xdef.XDFactory;
+import org.xdef.XDPool;
 import org.xdef.sys.ArrayReporter;
 
 public class X {
 	public static void main(String[] args) {
-		System.out.println("Xdefinition version: " + XDFactory.getXDVersion());
 		XDDocument xd = XDFactory.compileXD(null,
 "<xd:def xmlns:xd='" + XDConstants.XDEF42_NS_URI + "' root='A'>\n" +
 "  <xd:declaration>\n" +
@@ -26,5 +26,18 @@ public class X {
 		xd.xparse("<A><B a='USD'/><B a='CZK'/></A>", reporter);
 		System.out.println((!"USD\nCZK\n".equals(swr.toString()) || reporter.errorWarnings())
 			? "Error: " + reporter + ";\n" + swr : "OK");
+
+		System.out.println("Xdefinition version: " + XDFactory.getXDVersion());
+		System.setProperty(XDConstants.XDPROPERTY_STRING_CODE, "Windows-1250");
+		XDPool xp = XDFactory.compileXD(null,
+"<xd:def xmlns:xd='" + XDConstants.XDEF42_NS_URI + "' root='A'>\n" +
+"  <A><B xd:script='*;' a='string();'/></A>\n" +
+"</xd:def>");
+		reporter.clear();
+		xp.createXDDocument().xparse("<A><B a='ěščřžýáíéúů'/><B a='ÁÚŘ'/></A>", null);
+		System.out.println((reporter.errors() ? reporter.printToString() : "OK"));
+		reporter.clear();
+		xp.createXDDocument().xparse("<A><B a='والنشر مزيد منت'/><B a='Таблица аски'/></A>",reporter);
+		System.out.println((reporter.errors() ? reporter.printToString() : "OK"));
 	}
 }